Merge hotfix into main version 1.0.37 (#414)
* Fix OOM issue appearing on iPhone XR

This could potentially be an issue on other devices as well, since
we are trying to load a large binary into memory to try to parse
it as json. In the case of LCP this isn't what we want to do.

* Rebase off of main (#411)

* Update check-build-number.yml (#410)

What's this do?
Updates workflows

Why are we doing this? (w/ Notion link if applicable)
To support pushing TestFlight builds for merges into the hotfix branch

* Update upload-on-merge.yml (#409)

* Update build #

* Update project.pbxproj (#413)

Bump marketing version to create build

* Update project.pbxproj

---------

Co-authored-by: Jonathan Green <jonathan@razzard.com>